casting

+++ getting_started/arrays.c
@@ -5,7 +5,7 @@ int main() {
 
     // Since the available kit numbers array stores the same type across the board, the same amount of bytes are allocated 
     // so we can simple divide the size of the array by the size of one of the elements to get how many ACTUAL elements are in the array
-    int numberOfKits = sizeof(availableKitNumbers) / sizeof(availableKitNumbers[0]);
+    int numberOfKits = (int)(sizeof(availableKitNumbers) / sizeof(availableKitNumbers[0]));
 
     printf("We have %d kit numbers left.\n", numberOfKits);
